requirements.txt
Flask==2.3.3
flask-cors==4.0.0
gunicorn==23.0.0
openai==1.95.1
opencv-python==4.11.0.86
pandas==2.3.0
numpy==1.26.4
Pillow==11.2.1
requests==2.32.5
tensorflow==2.15.0
h5py>=3.8,<4.0
gradio